Rename AppSettingsBase to SharedPreferencesPropertyBackend
