+++ app/src/main/java/net/gsantner/opoc/preference/GsPreferenceFragmentCompat.java
@@ -59,7 +59,6 @@ import android.support.v7.preference.PreferenceScreen;
 import android.text.TextUtils;
 import android.view.View;
 
-import net.gsantner.opoc.util.AppSettingsBase;
 import net.gsantner.opoc.util.Callback;
 import net.gsantner.opoc.util.ContextUtils;
 
@@ -83,7 +82,7 @@ public abstract class GsPreferenceFragmentCompat extends PreferenceFragmentCompa
 
     public abstract String getFragmentTag();
 
-    protected abstract AppSettingsBase getAppSettings(Context context);
+    protected abstract SharedPreferencesPropertyBackend getAppSettings(Context context);
 
     //
     // Virtual
@@ -126,7 +125,7 @@ public abstract class GsPreferenceFragmentCompat extends PreferenceFragmentCompa
     //
 
     private final List<PreferenceScreen> _prefScreenBackstack = new ArrayList<>();
-    private AppSettingsBase _asb;
+    private SharedPreferencesPropertyBackend _asb;
     protected ContextUtils _cu;
 
     @Override
