+++ app/src/main/java/net/gsantner/opoc/preference/PropertyBackend.java
@@ -8,11 +8,11 @@
  *  https://www.apache.org/licenses/LICENSE-2.0
  *
 #########################################################*/
-package net.gsantner.opoc.util;
+package net.gsantner.opoc.preference;
 
 import java.util.List;
 
-@SuppressWarnings({"UnusedReturnValue", "SpellCheckingInspection"})
+@SuppressWarnings({"UnusedReturnValue", "SpellCheckingInspection", "unused", "SameParameterValue"})
 public interface PropertyBackend<TKEY, TTHIS> {
     String getString(TKEY key, String defaultValue);
